% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linktest.R
\name{convertToSeconds}
\alias{convertToSeconds}
\title{Trying to satisfy the documentation}
\usage{
convertToSeconds(y, m, d, hh, mm, ss, tz)
}
\arguments{
\item{y}{y}

\item{m}{m}

\item{d}{d}

\item{hh}{hh}

\item{mm}{mm}

\item{ss}{ss}

\item{tz}{tz}
}
\description{
Trying to satisfy the documentation
}
